requirements.txt
streamlit
PyPDF2
langchain
python-dotenv
openai
tiktoken
faiss-cpu
langchain-community
PyMuPDF
